refactor: collapse each SAM family into one module
Each of sam, sam2 and efficientsam kept its encoder and decoder as
one-function modules whose only caller was a forwarding method on the
model class, so reading how one family turns an image into a mask meant
bouncing across four files. Inline them into the family's model module,
matching the shape sam3 and yoloworld already have.

Move the grayscale/RGBA guard, copy-pasted into all three encoders, up
into the shared base so it is written once. Output masks are unchanged.
