+++ osam/_models/sam/__init__.py
@@ -2,3 +2,4 @@ from ._models import Sam100m  # noqa: F401
 from ._models import Sam300m  # noqa: F401
 from ._models import Sam600m  # noqa: F401
 from ._models import SamBase  # noqa: F401
+from ._models import get_input_size  # noqa: F401
