+++ osam/_models/sam2/_models.py
@@ -1,11 +1,12 @@
+from typing import cast
+
+import imgviz
 import numpy as np
 import numpy.typing as npt
 
 from osam import types
 from osam._models.sam import SamBase
-
-from . import _decoding
-from . import _encoding
+from osam._models.sam import get_input_size
 
 
 class Sam2(SamBase):
@@ -16,9 +17,36 @@ class Sam2(SamBase):
         license_spdx="Apache-2.0",
     )
 
-    def encode_image(self, image: npt.NDArray[np.uint8]) -> types.ImageEmbedding:
-        return _encoding.compute_image_embedding_from_image(
-            encoder_session=self._inference_sessions["encoder"], image=image
+    def _encode_image(self, image: npt.NDArray[np.uint8]) -> types.ImageEmbedding:
+        encoder_session = self._inference_sessions["encoder"]
+
+        input_height: int
+        input_width: int
+        input_height, input_width = encoder_session.get_inputs()[0].shape[2:]
+
+        input_: npt.NDArray[np.float32]
+        input_ = (
+            imgviz.resize(image, width=input_width, height=input_height).astype(
+                np.float32
+            )
+            / 255
+        )
+        #
+        input_ = input_ - np.array([0.485, 0.456, 0.406], dtype=np.float32) / np.array(
+            [0.229, 0.224, 0.225], dtype=np.float32
+        )
+        input_ = input_.transpose(2, 0, 1)[None]
+
+        outputs = encoder_session.run(output_names=None, input_feed={"input": input_})
+        image_embedding = cast(npt.NDArray[np.float32], outputs[0])
+        high_res_features1 = cast(npt.NDArray[np.float32], outputs[1])
+        high_res_features2 = cast(npt.NDArray[np.float32], outputs[2])
+
+        return types.ImageEmbedding(
+            original_height=image.shape[0],
+            original_width=image.shape[1],
+            embedding=image_embedding[0],
+            extra_features=[high_res_features1[0], high_res_features2[0]],
         )
 
     def _generate_mask_from_image_embedding(
@@ -26,14 +54,45 @@ class Sam2(SamBase):
         image_embedding: types.ImageEmbedding,
         prompt: types.Prompt,
     ) -> npt.NDArray[np.bool_]:
-        return _decoding.generate_mask_from_image_embedding(
-            decoder_session=self._inference_sessions["decoder"],
-            image_embedding=image_embedding,
-            prompt=prompt,
-            input_size=_encoding.get_input_size(
-                encoder_session=self._inference_sessions["encoder"]
-            ),
+        input_size: int = get_input_size(
+            encoder_session=self._inference_sessions["encoder"]
         )
+
+        input_point: npt.NDArray[np.float32] = np.array(prompt.points, dtype=np.float32)
+        input_point = input_point / np.array(
+            [
+                image_embedding.original_width / input_size,
+                image_embedding.original_height / input_size,
+            ],
+            dtype=np.float32,
+        )
+        input_label: npt.NDArray[np.float32] = np.array(
+            prompt.point_labels, dtype=np.float32
+        )
+
+        decoder_inputs = {
+            "image_embeddings": image_embedding.embedding[None],
+            "high_res_features1": image_embedding.extra_features[0][None],
+            "high_res_features2": image_embedding.extra_features[1][None],
+            "point_coords": input_point[None],
+            "point_labels": input_label[None],
+            "mask_input": np.zeros((1, 1, 256, 256), dtype=np.float32),
+            "has_mask_input": np.array([0], dtype=np.float32),
+            "orig_im_size": np.array(
+                (image_embedding.original_height, image_embedding.original_width),
+                dtype=np.int64,
+            ),
+        }
+        masks, scores, _low_res_mask = self._inference_sessions["decoder"].run(
+            None, decoder_inputs
+        )
+        masks = cast(npt.NDArray[np.float32], masks)
+        scores = cast(npt.NDArray[np.float32], scores)
+
+        mask: npt.NDArray[np.bool_] = (
+            masks[0, np.argmax(scores)] > 0.0
+        )  # (1, N, H, W) -> (H, W)
+        return mask
 
 
 class Sam2Tiny(Sam2):
